+++ Projet/Projet/Item.cs
@@ -13,9 +13,9 @@ namespace AtelierXNA
         public string CatégoriePersonnage { get; private set; }
         public string Nom { get; private set; }
         public int NiveauRequis { get; private set; }
-        public int Rareté {get; private set; }
+        public double Rareté {get; private set; }
 
-        public Item(int numeroID, string catégoriePersonnage, string nom, int niveauRequis, int rareté)
+        public Item(int numeroID, string catégoriePersonnage, string nom, int niveauRequis, double rareté)
         {
             Nom = nom;
             NiveauRequis = niveauRequis;
